fix: Rename ViewPagerAndroid to ViewPager
