+++ js/ViewPagerNativeComponent.js
